`Refactor App.js to use custom hooks and remove redundant state management`
